ecommerce/requirements.txt
alembic==1.17.2
blinker==1.9.0
boto3==1.42.30
botocore==1.42.30
click==8.3.1
Deprecated==1.3.1
Flask==3.1.2
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
limits==5.6.0
Mako==1.3.10
MarkupSafe==3.0.3
ordered-set==4.1.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.4
wrapt==2.0.1
